Added the Elevation Tile
When I created the elevation tile I had to manually change the
mapMaxZoom in the overlay_getElevationTileURL(bounds) function.

I suggest that we access that limit from the individual openlayers.html
files for each tile instead of hardcoding it at the top of index.html.
That will resolve the problem of tiles having different zoom limits.